backend/src/requirements.txt
django-ninja-jwt==5.2.10
django==5.0.1
django-ninja==1.1.0
django-ninja-extra==0.20.2
django-cors-headers==4.3.1
cryptography==41.0.7